AutoGenDemo/autogen-project/requirements.txt
autogen-agentchat
autogen-ext[openai,azure]
openai>=1.0.0
streamlit>=1.28.0
requests>=2.31.0
pandas>=2.0.0
plotly>=5.15.0
asyncio
python-dotenv
